feat: スキーマ共有用 @nonda/db パッケージを追加し web に Auth.js を導入
- packages/db を新設して apps/api/src/db/schema.ts を移動、
  apps/api からは re-export で参照
- apps/web に next-auth@5 beta + @auth/drizzle-adapter +
  drizzle-orm + @neondatabase/serverless を追加
- 空の packages/types を削除し README/CLAUDE.md を更新
