+++ apps/api/src/db/schema.ts
@@ -1,147 +1,1 @@
-import {
-  boolean,
-  index,
-  integer,
-  pgEnum,
-  pgTable,
-  primaryKey,
-  text,
-  time,
-  timestamp,
-  unique,
-  uuid,
-  varchar,
-} from "drizzle-orm/pg-core";
-
-export const timingEnum = pgEnum("timing_enum", ["morning", "afternoon", "evening"]);
-
-export const users = pgTable("users", {
-  id: uuid("id").primaryKey().defaultRandom(),
-  name: varchar("name", { length: 255 }),
-  email: varchar("email", { length: 255 }).notNull().unique(),
-  emailVerified: timestamp("email_verified", { withTimezone: true }),
-  image: text("image"),
-  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
-  updatedAt: timestamp("updated_at", { withTimezone: true })
-    .notNull()
-    .defaultNow()
-    .$onUpdate(() => new Date()),
-});
-
-export const accounts = pgTable(
-  "accounts",
-  {
-    userId: uuid("user_id")
-      .notNull()
-      .references(() => users.id, { onDelete: "cascade" }),
-    type: varchar("type", { length: 255 }).notNull(),
-    provider: varchar("provider", { length: 255 }).notNull(),
-    providerAccountId: varchar("provider_account_id", { length: 255 }).notNull(),
-    refresh_token: text("refresh_token"),
-    access_token: text("access_token"),
-    expires_at: integer("expires_at"),
-    token_type: varchar("token_type", { length: 255 }),
-    scope: varchar("scope", { length: 255 }),
-    id_token: text("id_token"),
-    session_state: varchar("session_state", { length: 255 }),
-  },
-  (table) => [
-    primaryKey({ columns: [table.provider, table.providerAccountId] }),
-    index("idx_accounts_user_id").on(table.userId),
-  ],
-);
-
-export const sessions = pgTable("sessions", {
-  sessionToken: varchar("session_token", { length: 255 }).primaryKey(),
-  userId: uuid("user_id")
-    .notNull()
-    .references(() => users.id, { onDelete: "cascade" }),
-  expires: timestamp("expires", { withTimezone: true }).notNull(),
-});
-
-export const verificationTokens = pgTable(
-  "verification_tokens",
-  {
-    identifier: varchar("identifier", { length: 255 }).notNull(),
-    token: varchar("token", { length: 255 }).notNull(),
-    expires: timestamp("expires", { withTimezone: true }).notNull(),
-  },
-  (table) => [primaryKey({ columns: [table.identifier, table.token] })],
-);
-
-export const medicines = pgTable(
-  "medicines",
-  {
-    id: uuid("id").primaryKey().defaultRandom(),
-    userId: uuid("user_id")
-      .notNull()
-      .references(() => users.id, { onDelete: "cascade" }),
-    name: varchar("name", { length: 255 }).notNull(),
-    photoUrl: text("photo_url"),
-    createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
-    updatedAt: timestamp("updated_at", { withTimezone: true })
-      .notNull()
-      .defaultNow()
-      .$onUpdate(() => new Date()),
-  },
-  (table) => [index("idx_medicines_user_id").on(table.userId)],
-);
-
-export const medicineTimings = pgTable(
-  "medicine_timings",
-  {
-    id: uuid("id").primaryKey().defaultRandom(),
-    medicineId: uuid("medicine_id")
-      .notNull()
-      .references(() => medicines.id, { onDelete: "cascade" }),
-    timing: timingEnum("timing").notNull(),
-    createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
-    updatedAt: timestamp("updated_at", { withTimezone: true })
-      .notNull()
-      .defaultNow()
-      .$onUpdate(() => new Date()),
-  },
-  (table) => [unique("uq_medicine_timings_medicine_id_timing").on(table.medicineId, table.timing)],
-);
-
-export const medicationLogs = pgTable(
-  "medication_logs",
-  {
-    id: uuid("id").primaryKey().defaultRandom(),
-    medicineId: uuid("medicine_id")
-      .notNull()
-      .references(() => medicines.id, { onDelete: "cascade" }),
-    timing: timingEnum("timing").notNull(),
-    recordedAt: timestamp("recorded_at", { withTimezone: true }).notNull(),
-    isTaken: boolean("is_taken").notNull().default(false),
-    createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
-    updatedAt: timestamp("updated_at", { withTimezone: true })
-      .notNull()
-      .defaultNow()
-      .$onUpdate(() => new Date()),
-  },
-  (table) => [
-    index("idx_medication_logs_medicine_id").on(table.medicineId),
-    index("idx_medication_logs_recorded_at").on(table.recordedAt),
-    index("idx_medication_logs_medicine_recorded_at").on(table.medicineId, table.recordedAt),
-  ],
-);
-
-export const notificationSettings = pgTable(
-  "notification_settings",
-  {
-    id: uuid("id").primaryKey().defaultRandom(),
-    userId: uuid("user_id")
-      .notNull()
-      .references(() => users.id, { onDelete: "cascade" }),
-    timing: timingEnum("timing").notNull(),
-    notifyTime: time("notify_time").notNull(),
-    isEnabled: boolean("is_enabled").notNull().default(true),
-    createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
-    updatedAt: timestamp("updated_at", { withTimezone: true })
-      .notNull()
-      .defaultNow()
-      .$onUpdate(() => new Date()),
-  },
-  (table) => [unique("uq_notification_settings_user_id_timing").on(table.userId, table.timing)],
-);
+export * from "@nonda/db";
