+++ packages/db/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./schema.js";
